fix(server): SeaORM column mapping, C++→Rust migration, and stale-job resubmit fix
This branch represents the migration of the server from the legacy C++
implementation to the Rust+SeaORM stack, the API compatibility work
required to keep the C++-shaped client working against it, the
operational hardening needed to make the resulting binary deployable
in this environment, and the stale-job resubmit fix that prevents
failed jobs from looping back into the submission queue on every
reconnect.

Database & entity mapping:
- Correct SeaORM entity column mappings to match the live MySQL
  schema (snake_case columns, widened `jobserver_filedownload.user`
  and `job` from INT to BIGINT, etc.) — the Rust entities were
  written for an older schema shape and the column-name mismatch was
  causing every DB-over-WS request to fail
- Add the missing live table coverage for the SeaORM entities
  (`jobserver_clusterjob`, `jobserver_clusterjobstatus`,
  `jobserver_filelistcache`, `jobserver_filedownload`,
  `jobserver_clusteruuid`)

API compatibility:
- Restore the C++-compatible API paths for file endpoints
  (`/file/apiv1/file/...` was the old shape; the C++ client still
  hits those paths)
- Make the JSON HTTP endpoints tolerant of missing `Content-Type`
  headers (the existing client doesn't always set them)

Operational hardening:
- Add comprehensive verbose logging throughout the application
- Add `WS: Received binary message` / `WS: Parsed message` style
  logging to file download / upload handlers so the next protocol
  mismatch is diagnosable from a single `docker logs`
- Silence the `tungstenite`/`russh`/`sqlx` dependency log noise so
  the app's own logs aren't drowned out
- `chore(logging): line-buffer server log output` — wrap the logfile
  tee in `stdbuf -oL -eL` so `docker logs` and `/app/logs/logfile`
  show output immediately rather than only on buffer flush
- Set `RUST_LOG` default to `info` so non-debug logs are not silent
  on a fresh deploy
- Add a `tracing::debug!` around the UUID-token DB lookup in
  `cluster::manager::handle_new_connection` so future `Invalid
  token` failures are diagnosable

Stale-job resubmit fix:
The server's `check_resend_jobs` only ordered by `timestamp DESC`, so
a `Pending` and `Submitting` history row inserted in the same
millisecond at job-creation time resolved non-deterministically,
allowing failed jobs whose `Submitting` row happened to sort first
to re-fire on every reconnect. Fixed in two parts:
- Added `order_by_desc(job_history::Column::Id)` as a deterministic
  tie-breaker (the auto-increment id is always strictly increasing,
  so later-inserted rows sort first under same-ms ties)
- Added a `seen_terminal_or_inflight` set: while building
  `latest_per_job`, any job whose history contains a row with state
  outside the resend set (i.e. `SUBMITTED`/`RUNNING`/`COMPLETED`/
  error) is flagged and skipped. This stops the resubmit loop from
  looping failed jobs back into "Resubmitting" when their same-ms
  `Submitting` row wins the tie.

Verified end-to-end against the local Slurm cluster: stale jobs
(7/11/12/13/14/15/16/17/18/19/20/21) no longer re-fire on every
reconnect, and the server handles fresh job submissions cleanly
through to `state 500`.

+++ src/src/db/entities/file_download.rs
@@ -6,8 +6,8 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
-    pub user: i32,
-    pub job: i32,
+    pub user: i64,
+    pub job: i64,
     pub cluster: String,
     pub bundle: String,
     pub uuid: String,
